requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
Django==5.2.8
django-rest-framework==0.1.0
djangorestframework==3.16.1
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.121.3
fastapi-cli==0.0.16
fastapi-cloud-cli==0.5.0
fastar==0.6.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==1.0.7
langchain-ollama==1.0.0
langsmith==0.4.44
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ollama==0.6.1
orjson==3.11.4
packaging==25.0
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
python-dotenv==1.2.1
python-multipart==0.0.20
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rich-toolkit==0.16.0
rignore==0.7.6
sentry-sdk==2.45.0
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.44
sqlparse==0.5.3
starlette==0.50.0
tenacity==9.1.2
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
zstandard==0.25.0
